src/main/java/uk/gov/hmcts/reform/iacaseapi/domain/handlers/presubmit/UpdateHmcResponseEventHandler.java
package uk.gov.hmcts.reform.iacaseapi.domain.handlers.presubmit;


import org.springframework.stereotype.Component;
import uk.gov.hmcts.reform.iacaseapi.domain.entities.AsylumCase;
import uk.gov.hmcts.reform.iacaseapi.domain.entities.AsylumCaseFieldDefinition;
import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.Event;
import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.Callback;
import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubmitCallbackResponse;
import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubmitCallbackStage;
import uk.gov.hmcts.reform.iacaseapi.domain.handlers.PreSubmitCallbackHandler;
import uk.gov.hmcts.reform.iacaseapi.domain.handlers.PreSubmitCallbackStateHandler;

import static java.util.Objects.requireNonNull;
import static uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.field.YesOrNo.YES;
@Component
public class UpdateHmcResponseEventHandler implements PreSubmitCallbackHandler<AsylumCase> {
    public boolean canHandle(
            PreSubmitCallbackStage callbackStage,
            Callback<AsylumCase> callback
    ) {
        requireNonNull(callbackStage, "callbackStage must not be null");
        requireNonNull(callback, "callback must not be null");

        return callback.getEvent() == Event.UPDATE_HMC_RESPONSE;
    }


    public PreSubmitCallbackResponse<AsylumCase> handle(
            PreSubmitCallbackStage callbackStage,
            Callback<AsylumCase> callback
    ) {
        if (!canHandle(callbackStage, callback)) {
            throw new IllegalStateException("Cannot handle callback");
        }

        final AsylumCase asylumCase =
                callback
                        .getCaseDetails()
                        .getCaseData();

        boolean isIntegrated = asylumCase.read(AsylumCaseFieldDefinition.IS_INTEGRATED)
                .map(yesOrNo -> yesOrNo.equals(YES))
                .orElse(false);

        if(!isIntegrated){
            asylumCase.write(AsylumCaseFieldDefinition.IS_INTEGRATED, YES);
        }
        return new PreSubmitCallbackResponse<>(asylumCase);
    }
}
